Move React specific code to its own module
