+++ package.json
@@ -19,7 +19,10 @@
   "main": "dist/index.js",
   "module": "dist/index.esm.js",
   "unpkg": "dist/foo.umd.js",
-  "exports": "./dist/foo.modern.js",
+  "exports": {
+    ".": "./dist/index.esm.js",
+    "./react": "./dist/react.esm.js"
+  },
   "types": "dist/index.d.ts",
   "files": [
     "**"
@@ -27,8 +30,8 @@
   "license": "MIT",
   "scripts": {
     "clean": "rimraf dist",
-    "dev": "microbundle watch",
-    "build": "microbundle",
+    "dev": "microbundle watch src/*.ts --external react",
+    "build": "microbundle build src/*.ts --external react",
     "test": "jest",
     "docs": "typedoc src/index.ts",
     "prepublishOnly": "yarn test && yarn build"
